refactor: deduplicate ROT13/Base64 obfuscation into shared utility
The obfuscation logic existed twice (app_database.dart and
import_parser.dart) plus a third call site in the AI generator using
the parser's copy. Move encode/decode into
lib/core/utils/obfuscation.dart and use it everywhere.

Also document in docs/import-export/format.md that the encoding is
pure spoiler protection, not a security feature: it is trivially
reversible and resolution notes travel decodably with every shared
export.

Closes #117
