+++ lib/core/database/app_database.dart
@@ -5,6 +5,8 @@ import 'package:drift/native.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:path/path.dart' as p;
 
+import '../utils/obfuscation.dart';
+
 part 'app_database.g.dart';
 
 // --- Tables ---
@@ -344,7 +346,7 @@ class AppDatabase extends _$AppDatabase {
         if (effectiveUnit != null && effectiveUnit.isNotEmpty)
           'unit': effectiveUnit,
         if (v.resolution != null)
-          'resolution': _obfuscateResolution({
+          'resolution': obfuscateResolution({
             'outcome': v.resolution!.outcome,
             if (v.resolution!.numericOutcome != null)
               'numericOutcome': v.resolution!.numericOutcome,
@@ -378,7 +380,7 @@ class AppDatabase extends _$AppDatabase {
         'hasKnownAnswer': q.hasKnownAnswer,
         if (q.knownAnswer != null) 'knownAnswer': q.knownAnswer,
         if (effectiveUnit != null && effectiveUnit.isNotEmpty) 'unit': effectiveUnit,
-        'resolution': _obfuscateResolution({
+        'resolution': obfuscateResolution({
           'outcome': v.resolution!.outcome,
           if (v.resolution!.numericOutcome != null)
             'numericOutcome': v.resolution!.numericOutcome,
@@ -545,7 +547,7 @@ class AppDatabase extends _$AppDatabase {
             'createdAt': estimate.createdAt.toIso8601String(),
           },
         if (resolution != null)
-          'resolution': _obfuscateResolution({
+          'resolution': obfuscateResolution({
             'outcome': resolution.outcome,
             'numericOutcome': resolution.numericOutcome,
             'notes': resolution.notes,
@@ -559,21 +561,6 @@ class AppDatabase extends _$AppDatabase {
       'questions': result,
     };
   }
-}
-
-/// ROT13 anwenden, dann Base64 kodieren.
-String _obfuscateResolution(Map<String, dynamic> resolution) {
-  final plain = jsonEncode(resolution);
-  final rot13 = _rot13(plain);
-  return base64Encode(utf8.encode(rot13));
-}
-
-String _rot13(String input) {
-  return String.fromCharCodes(input.codeUnits.map((c) {
-    if (c >= 65 && c <= 90) return (c - 65 + 13) % 26 + 65;
-    if (c >= 97 && c <= 122) return (c - 97 + 13) % 26 + 97;
-    return c;
-  }));
 }
 
 LazyDatabase _openConnection() {
